Version 1.0.01
Fixed EnvSub's date-time math so that if a newer time is subtracted from an older one, the result will be reported correctly as a negative number rather than a large positive number.

Fixed a rare problem with naked modifier keys used as hotkeys. For example, if RControl (by itself) was a hotkey, sometimes the control key would get stuck down after a Send operation that involved the control key.

Fixed syntax error reporting: in rare cases such as Else If statements, the line above where the syntax error actually occurred was reported.

Added the built-in variable %0%, which contains the number of command line parameters that were passed into the script. This also improves compatibility with AutoIt2. [thanks BoBo]

Added command BlockInput, which prevents the user from doing anything with the keyboard, mouse, and maybe other input devices as well. [thanks Jonathan Bennett & beardboy]

+++ Source/util.cpp
@@ -135,12 +135,15 @@ __int64 FileTimeSecondsUntil(FILETIME *pftStart, FILETIME *pftEnd)
 	if (!pftStart || !pftEnd) return 0;
 
 	// The calculation is done this way for compilers that don't support 64-bit math operations (not sure which):
+	// Note: This must be LARGE vs. ULARGE because we want the calculation to be signed for cases where
+	// pftStart is greater than pftEnd:
 	ULARGE_INTEGER uiStart, uiEnd;
 	uiStart.LowPart = pftStart->dwLowDateTime;
 	uiStart.HighPart = pftStart->dwHighDateTime;
 	uiEnd.LowPart = pftEnd->dwLowDateTime;
 	uiEnd.HighPart = pftEnd->dwHighDateTime;
-	return (__int64)((uiEnd.QuadPart - uiStart.QuadPart) / 10000000); // Convert from tenths-of-microsecond.
+	// Must do at least the inner cast to avoid losing negative results:
+	return (__int64)((__int64)(uiEnd.QuadPart - uiStart.QuadPart) / 10000000); // Convert from tenths-of-microsecond.
 }
